refactor: unify marketplace & landing header
